Rename nfpm.field_set -> nfpm.field_sets
