+++ src/python/pants/backend/experimental/nfpm/register.py
@@ -1,7 +1,7 @@
 # Copyright 2023 Pants project contributors (see CONTRIBUTORS.md).
 # Licensed under the Apache License, Version 2.0 (see LICENSE).
 
-from pants.backend.nfpm import field_set, generate_config_rules
+from pants.backend.nfpm import field_sets, generate_config_rules
 from pants.backend.nfpm import rules as nfpm_rules
 from pants.backend.nfpm.target_types import (
     NfpmApkPackage,
@@ -22,7 +22,7 @@ def target_types():
 
 def rules():
     return [
-        *field_set.rules(),
+        *field_sets.rules(),
         *generate_config_rules.rules(),
         *nfpm_rules.rules(),
     ]
