+++ src/python/pants/backend/nfpm/rules.py
@@ -1,6 +1,8 @@
 # Copyright 2023 Pants project contributors (see CONTRIBUTORS.md).
 # Licensed under the Apache License, Version 2.0 (see LICENSE).
 
+from __future__ import annotations
+
 from pants.engine.rules import collect_rules
